fastapi_backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
supabase
python-jose[cryptography]
passlib[bcrypt]